refactor(userspace-dp): split queue_service tests.rs by concern
The single userspace-dp/src/afxdp/cos/queue_service/tests.rs had grown
to 4384 lines / 85 tests, well past the modularity-discipline threshold
that the rest of the crate holds to (codex-review-174 #4, tracked in
#4665). Split it into a tests/ subdirectory of per-concern sibling
submodules loaded from tests/mod.rs, mirroring the event_stream/codec
split convention (#4651):

  selector  (31)  guarantee/surplus phase selection, RR cursors,
                  priority + weighted sharing, promotion pairing
  drain     (12)  drain_shaped_tx priming/service, TX-progress gating,
                  exact local/prepared FIFO scratch commit/settle/release
  wakeup     (8)  wakeup-tick estimation, runnable restore, DSCP rewrite
  waterfill (18)  #1614 waterfill selector modes + Phase-1 honored-set
                  admission accounting
  sojourn    (6)  #1829 committed-prefix sojourn sampling
  refund     (5)  Phase-1 honor refund + shared-lease guaranteed refill
  submit     (5)  submit_local/prepared owner tx-counter bumps +
                  flow-fair local-scratch offset recycling

This is pure test-code motion. Every test and helper body was moved
byte-identical (extracted by exact line ranges, not retyped); no test
logic, name, or assertion changed. The shared header imports, the epoch
constant TEST_EPOCH_DURATION_NS, a mid-file `use` that sat beside the
drain tests, and the one cross-concern fixture
waterfill_guarantee_rate_root (used by both waterfill and refund) live
in tests/mod.rs and reach each submodule via `use super::*` — the
transitive glob and explicit-use re-export through the parent module
were verified to resolve before the split.

The only production-file change is the test-module wiring: the hot
selectors in mod.rs are untouched; `#[path = "tests.rs"] mod tests;`
becomes plain `mod tests;`, which resolves to the new tests/mod.rs.

Validation:
  - Code-line multiset of the split output is identical to the
    pre-split tests.rs (no line lost, altered, or duplicated).
  - Crate-wide #[test]/#[tokio::test] count unchanged: 3843 == 3843;
    queue_service 85 == 85.
  - cargo build (release) OK.
  - FULL serial suite green: main test binary 3763 passed / 0 failed /
    2 ignored; 60 + 8 + 22 + 1 in the other binaries.
  - `cargo test queue_service` runs all 85 across the 7 submodules,
    all passing.

Closes #4665
